requirements.txt
# Core Django
Django==5.1.4
djangorestframework==3.15.2

# Database
psycopg2-binary==2.9.9

# Static Files & Media
whitenoise==6.8.2
Pillow==10.4.0

# Forms & UI
django-crispy-forms==2.3
crispy-bootstrap5==2024.10

# Rich Text Editor
django-ckeditor==6.7.1

# Filtering
django-filter==24.3

# Environment Configuration
django-environ==0.11.2

# Email
django-ses==4.1.1

# Cloud Storage (AWS)
django-storages==1.14.4
boto3==1.35.69

# Web Server
gunicorn==23.0.0

# Machine Learning
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2

# NLP Libraries
nltk==3.9.1
mpstemmer @ git+https://github.com/ariaghora/mpstemmer.git
PySastrawi==1.2.0

# Development & Testing
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==6.0.0
black==24.10.0
flake8==7.1.1
isort==5.13.2

# Security
django-cors-headers==4.6.0
django-ratelimit==4.1.0

# Monitoring & Logging
sentry-sdk==2.18.0

# Task Queue (optional for async tasks)
celery==5.4.0
redis==5.2.0

# API Documentation
drf-spectacular==0.27.2